Renaming DocsLayout to SidebarSidecarLayout
