+++ src/layouts/sidebar-sidecar/utils/_local_platform-docs-mdx/style.module.css
