+++ src/layouts/sidebar-sidecar/utils/prepare-nav-data-for-client.ts
