+++ src/layouts/sidebar-sidecar/utils/product-url-adjusters.ts
