--- src/pages/_layout-dev/style.module.css
@@ -1,48 +1,0 @@
-/* DEMO CONTENT */
-
-.headerDemo {
-  border: 2px solid black;
-  background: rgba(128, 128, 128, 1);
-  height: var(--sticky-bars-height);
-}
-
-.sidebarDemo {
-  border: 2px solid blue;
-  background: rgba(0, 0, 255, 0.4);
-
-  &.isLong {
-    min-height: 2000px;
-  }
-}
-
-.mainDemo {
-  border: 3px solid green;
-  background: rgba(0, 255, 0, 0.4);
-  /* demo meant to match potential docs layout use...
-     the 50vh bottom padding ensures that all
-     jump-nav items have room to be jumped to */
-  padding-bottom: 50vh;
-
-  &.isLong {
-    min-height: 150vh;
-  }
-}
-
-.sidecarDemo {
-  border: 2px solid blue;
-  background: rgba(0, 0, 255, 0.4);
-
-  &.isLong {
-    min-height: 2000px;
-  }
-}
-
-.footerDemo {
-  border: 2px solid black;
-  background: rgba(128, 128, 128, 1);
-  height: 100px;
-
-  &.isLong {
-    height: 900px;
-  }
-}
