+++ src/pages/vault/api-docs/[[...page]].tsx
@@ -1,10 +1,10 @@
 import { ReactElement } from 'react'
 import vaultData from 'data/vault.json'
 import { Product } from 'types/products'
-import { getStaticGenerationFunctions } from 'layouts/docs/server'
-import DocsLayout from 'layouts/docs'
+import { getStaticGenerationFunctions } from 'layouts/sidebar-sidecar/server'
+import SidebarSidecarLayout from 'layouts/sidebar-sidecar'
 import DocsView from 'views/docs-view'
-import { vaultUrlAdjuster } from 'layouts/docs/utils/product-url-adjusters'
+import { vaultUrlAdjuster } from 'layouts/sidebar-sidecar/utils/product-url-adjusters'
 
 const basePath = 'api-docs'
 const baseName = 'API'
@@ -22,7 +22,7 @@ const { getStaticPaths, getStaticProps } = getStaticGenerationFunctions({
   additionalRemarkPlugins: [vaultUrlAdjuster],
 })
 
-VaultApiDocsPage.layout = DocsLayout
+VaultApiDocsPage.layout = SidebarSidecarLayout
 
 export { getStaticPaths, getStaticProps }
 export default VaultApiDocsPage
