+++ src/pages/vault/docs/[[...page]].tsx
@@ -1,12 +1,12 @@
 import { ReactElement } from 'react'
 import vaultData from 'data/vault.json'
 import { Product } from 'types/products'
-import { getStaticGenerationFunctions } from 'layouts/docs/server'
-import DocsLayout from 'layouts/docs'
+import { getStaticGenerationFunctions } from 'layouts/sidebar-sidecar/server'
+import SidebarSidecarLayout from 'layouts/sidebar-sidecar'
 import DocsView from 'views/docs-view'
 import Columns from 'components/author-primitives/vault/columns'
 import InlineTag from 'components/author-primitives/vault/inline-tag'
-import { vaultUrlAdjuster } from 'layouts/docs/utils/product-url-adjusters'
+import { vaultUrlAdjuster } from 'layouts/sidebar-sidecar/utils/product-url-adjusters'
 
 const basePath = 'docs'
 const baseName = 'Docs'
@@ -28,7 +28,7 @@ const { getStaticPaths, getStaticProps } = getStaticGenerationFunctions({
   additionalRemarkPlugins: [vaultUrlAdjuster],
 })
 
-VaultDocsPage.layout = DocsLayout
+VaultDocsPage.layout = SidebarSidecarLayout
 
 export { getStaticPaths, getStaticProps }
 export default VaultDocsPage
